+++ TabletopTweaks/NewContent/FighterAdvancedWeaponTrainings/DefensiveWeaponTraining.cs
@@ -14,6 +14,7 @@ namespace TabletopTweaks.NewContent.FighterAdvancedWeaponTrainings {
             var DefensiveWeaponTrainingFeature = Helpers.CreateBlueprint<BlueprintFeature>("DefensiveWeaponTrainingFeature", bp => {
                 bp.IsClassFeature = true;
                 bp.Ranks = 1;
+                bp.Groups = new FeatureGroup[] { FeatureGroup.WeaponTraining };
                 bp.SetName("Defensive Weapon Training");
                 bp.SetDescription("The fighter gains a +1 shield bonus to his Armor Class. The fighter adds half his weapon’s enhancement bonus (if any) to this shield bonus. " +
                     "When his weapon training bonus for weapons from the associated fighter weapon group reaches +4, this shield bonus increases to +2. " +