+++ TabletopTweaks/NewContent/FighterAdvancedWeaponTrainings/TrainedThrow.cs
@@ -21,6 +21,7 @@ namespace TabletopTweaks.NewContent.FighterAdvancedWeaponTrainings {
             var TrainedThrowFeature = Helpers.CreateBlueprint<BlueprintFeature>("TrainedThrowFeature", bp => {
                 bp.IsClassFeature = true;
                 bp.Ranks = 1;
+                bp.Groups = new FeatureGroup[] { FeatureGroup.WeaponTraining };
                 bp.SetName("Trained Throw");
                 bp.SetDescription("When the fighter makes a ranged attack with a thrown weapon and applies his Dexterity modifier" +
                     " on attack rolls and his Strength modifier on damage rolls, he doubles his weapon training bonus on damage rolls.");